refactor: chart header right buttons 레이지 로딩
