+++ components/hospital/icu/main/chart/selected-chart/chart-header/header-right-buttons/delete-chart-dialog.tsx
@@ -1,8 +1,8 @@
+import LargeLoaderCircle from '@/components/common/large-loader-circle'
 import WarningMessage from '@/components/common/warning-message'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogDescription,
   DialogFooter,
@@ -10,17 +10,15 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog'
-import { toast } from '@/components/ui/use-toast'
-import {
-  deleteAllCharts,
-  deleteChart,
-  deleteOrders,
-} from '@/lib/services/icu/chart/delete-icu-chart'
-import { cn } from '@/lib/utils/utils'
-import { LoaderCircle, Trash2 } from 'lucide-react'
+import { Trash2 } from 'lucide-react'
+import dynamic from 'next/dynamic'
 import { useParams } from 'next/navigation'
 import { useState } from 'react'
 
+const LazyDeleteChartButtons = dynamic(() => import('./delete-chart-buttons'), {
+  ssr: false,
+  loading: () => <LargeLoaderCircle size={40} />,
+})
 export default function DeleteChartDialog({
   icuChartId,
   name,
@@ -34,36 +32,6 @@ export default function DeleteChartDialog({
 }) {
   const { target_date } = useParams()
   const [isDialogOpen, setIsDialogOpen] = useState(false)
-  const [isDeleting, setIsDeleting] = useState(false)
-  const [isDeletingAll, setIsDeletingAll] = useState(false)
-
-  const handleDeleteChart = async () => {
-    setIsDeleting(true)
-
-    // 첫 차트인 경우 오더만 삭제, 2일차이상의 경우 차트 전체 삭제
-    if (isFirstChart) {
-      await deleteOrders(icuChartId)
-    } else {
-      await deleteChart(icuChartId)
-    }
-
-    toast({
-      title: '차트가 삭제되었습니다',
-    })
-    setIsDeleting(false)
-    setIsDialogOpen(false)
-  }
-  const handleDeleteAllCharts = async () => {
-    setIsDeletingAll(true)
-
-    await deleteAllCharts(icuIoId)
-
-    toast({
-      title: `${name}의 모든차트가 삭제되었습니다`,
-    })
-    setIsDeletingAll(false)
-    setIsDialogOpen(false)
-  }
 
   return (
     <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
@@ -83,30 +51,12 @@ export default function DeleteChartDialog({
         </DialogHeader>
 
         <DialogFooter>
-          <DialogClose asChild>
-            <Button type="button" variant="outline">
-              취소
-            </Button>
-          </DialogClose>
-          <Button
-            onClick={handleDeleteChart}
-            disabled={isDeleting || isDeletingAll}
-          >
-            선택차트삭제
-            <LoaderCircle
-              className={cn(isDeleting ? 'ml-2 animate-spin' : 'hidden')}
-            />
-          </Button>
-          <Button
-            onClick={handleDeleteAllCharts}
-            variant="destructive"
-            disabled={isDeleting || isDeletingAll}
-          >
-            모든차트삭제
-            <LoaderCircle
-              className={cn(isDeletingAll ? 'ml-2 animate-spin' : 'hidden')}
-            />
-          </Button>
+          <LazyDeleteChartButtons
+            setIsDialogOpen={setIsDialogOpen}
+            isFirstChart={isFirstChart}
+            icuChartId={icuChartId}
+            icuIoId={icuIoId}
+          />
         </DialogFooter>
       </DialogContent>
     </Dialog>
