+++ components/hospital/icu/main/chart/selected-chart/chart-header/header-right-buttons/export-dialog/export-buttons.tsx
@@ -1,0 +1,26 @@
+import type { SelectedChart } from '@/types/icu/chart'
+import { Dispatch, SetStateAction } from 'react'
+import ExportPdfButton from './export-pdf-button'
+import ExportPngButton from './export-png-button'
+
+export default function ExportButtons({
+  chartData,
+  setIsDialogOpen,
+}: {
+  chartData: SelectedChart
+  setIsDialogOpen: Dispatch<SetStateAction<boolean>>
+}) {
+  return (
+    <>
+      <ExportPngButton
+        chartData={chartData}
+        setIsDialogOpen={setIsDialogOpen}
+      />
+
+      <ExportPdfButton
+        chartData={chartData}
+        setIsDialogOpen={setIsDialogOpen}
+      />
+    </>
+  )
+}
